+++ codegen/example/xata.ts
@@ -1,10 +1,4 @@
-import {
-  BaseClient,
-  Repository,
-  RestRespositoryFactory,
-  XataClientOptions,
-  XataRecord,
-} from "../../client/src";
+import { buildClient, BaseClientOptions, XataRecord } from '../../client/src';
 
 export interface Team {
   name?: string | null;
@@ -23,23 +17,17 @@ export interface User {
 
 export type UserRecord = User & XataRecord;
 
-const links = { teams: [["owner", "users"]], users: [["team", "teams"]] };
+export type DatabaseSchema = {
+  teams: Team;
+  users: User;
+};
 
-export class XataClient extends BaseClient<{
-  teams: Repository<Team>;
-  users: Repository<User>;
-}> {
-  constructor(options?: XataClientOptions) {
-    super(
-      { databaseURL: "https://test-r5vcv5.xata.sh/db/test", ...options },
-      links
-    );
+const links = { teams: [['owner', 'users']], users: [['team', 'teams']] };
 
-    const factory = options?.repositoryFactory || new RestRespositoryFactory();
+const DatabaseClient = buildClient<DatabaseSchema>();
 
-    this.db = {
-      teams: factory.createRepository(this, "teams", links),
-      users: factory.createRepository(this, "users", links),
-    };
+export class XataClient extends DatabaseClient {
+  constructor(options?: BaseClientOptions) {
+    super({ databaseURL: 'https://test-r5vcv5.xata.sh/db/test', ...options }, links);
   }
 }