+++ client/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @xata.io/client
 
+## 0.7.2
+
+### Patch Changes
+
+- 4803b6f: Memoize ApiClient namespaces
+- 1f268d7: Add search in XataClient
+- d58c3d9: Hide private helper utilities
+- f3b731b: Expose branch resolution API
+
 ## 0.7.1
 
 ### Patch Changes